+++ jgclark.DashboardReact/src/react/components/ProjectItem.jsx
@@ -2,14 +2,15 @@
 //--------------------------------------------------------------------------
 // Dashboard React component to show a Project's item
 // Called by ItemRow component
-// Last updated 18.6.2024 for v2.0.0-b9 by @jgclark
+// Last updated 24.6.2024 for v2.0.0-b14 by @jgclark
 //--------------------------------------------------------------------------
 
 import * as React from 'react'
-import { CircularProgressbar, CircularProgressbarWithChildren, buildStyles } from 'react-circular-progressbar'
-import 'react-circular-progressbar/dist/styles.css'
+// import { CircularProgressbar, CircularProgressbarWithChildren, buildStyles } from 'react-circular-progressbar'
+// import 'react-circular-progressbar/dist/styles.css'
 import type { TSectionItem } from '../../types.js'
 import { useAppContext } from './AppContext.jsx'
+import ProjectIcon from './ProjectIcon.jsx'
 import { getFolderFromFilename } from '@helpers/folders'
 import { logDebug, clo } from '@helpers/react/reactDev.js'
 import { extractModifierKeys } from '@helpers/react/reactMouseKeyboard.js'
@@ -25,7 +26,7 @@ function ProjectItem({ item }: Props): React.Node {
   const noteTitle = item.project?.title ?? '<no title>'
   const folderNamePart = sharedSettings?.includeFolderName && getFolderFromFilename(itemFilename) !== '' ? `${getFolderFromFilename(itemFilename)} / ` : ''
   // logDebug(`ProjectItem`, `for ${itemFilename} (${folderNamePart} / ${noteTitle})`)
-  const percentComplete = item.project?.percentComplete ?? 0
+  // const percentComplete = item.project?.percentComplete ?? 0
   // const percentCompleteStr = isNaN(percentComplete) ? '' : ` ${String(percentComplete)}%`
   const progressText = item.project?.lastProgressComment ?? ''
   const noteTitleWithOpenAction = (
@@ -33,43 +34,6 @@ function ProjectItem({ item }: Props): React.Node {
       <i className="fa-regular fa-file-lines pad-right"></i>
       {noteTitle}
     </a>
-  )
-
-  // const projectIcon = (percentComplete > 0) && (
-  //   <>
-  //     <div className="progress-bar">
-  //       <progress value={String(percentComplete)} min="0" max="100">
-  //         {percentCompleteStr}
-  //       </progress>
-  //     </div>
-  //   </>
-  // )
-  // using https://www.npmjs.com/package/react-circular-progressbar
-  const projectIcon = (
-    <CircularProgressbarWithChildren
-      // background path
-      /*text={`${percentage}%`}*/
-      value={percentComplete}
-      strokeWidth={50}
-      styles={buildStyles({
-        strokeLinecap: "butt",
-        backgroundColor: "var(--bg-sidebar-color)",
-        // backgroundColor: "transparent",
-        pathColor: "var(--tint-color)",
-      })}
-    >
-      {/* foreground path */}
-      <CircularProgressbar
-        value={100}
-        strokeWidth={5}
-        styles={buildStyles({
-          strokeLinecap: "butt",
-          backgroundColor: "var(--bg-sidebar-color)",
-          // backgroundColor: "transparent",
-          pathColor: "var(--tint-color)",
-        })}
-      ></CircularProgressbar>
-    </CircularProgressbarWithChildren>
   )
 
   const progressContent = progressText && (
@@ -109,9 +73,8 @@ function ProjectItem({ item }: Props): React.Node {
 
   return (
     <div className="sectionItemRow" id={item.ID}>
-      <div className="projectIcon">
-        {/* <i id={`${item.ID}I`} className="fa-regular fa-file-lines"></i> */}
-        {projectIcon}
+      <div>
+        <ProjectIcon item={item} />
       </div>
 
       <div className="sectionItemContent sectionItem">