+++ jgclark.Reviews/src/reviewHelpers.js
@@ -886,13 +886,13 @@ export class Project {
 
   /**
    * Returns title of note as folder name + link, also showing complete or cancelled where relevant.
-   * Supports 'Markdown' or 'HTML' styling.
-   * @param {string} style 'Markdown' or 'HTML'
+   * Supports 'Markdown' or 'HTML' styling or simpler 'list' styling
+   * @param {string} style 'Markdown' or 'HTML' or 'list'
    * @param {boolean} includeFolderName whether to include folder name at the start of the entry.
    * @return {string} - title as wikilink
    */
   decoratedProjectTitle(style: string, includeFolderName: boolean): string {
-    const folderNamePart = includeFolderName ? `${this.folder} ` : ''
+    const folderNamePart = includeFolderName ? `${this.folder} / ` : ''
     const titlePart = this.title ?? '(error, not available)'
     // const titlePartEncoded = encodeURIComponent(this.title) ?? '(error, not available)'
     switch (style) {
@@ -930,6 +930,18 @@ export class Project {
         }
       }
 
+      case 'list': {
+        if (this.isCompleted) {
+          return `${folderNamePart}[[${titlePart}]]`
+        } else if (this.isCancelled) {
+          return `~~${folderNamePart}[[${titlePart}]]~~`
+        } else if (this.isPaused) {
+          return `⏸ **Paused**: ${folderNamePart}[[${titlePart}]]`
+        } else {
+          return `${folderNamePart}[[${titlePart}]]` // if this has a [ ] prefix then it of course turns it into a task, which is probably not what we want.
+        }
+      }
+
       default:
         logWarn('Project::decoratedProjectTitle', `Unknown style '${style}'; nothing returned.`)
         return ''
@@ -937,7 +949,7 @@ export class Project {
   }
 
   /**
-   * Returns line showing more detailed summary of the project, for output in Rich (HTML) or Markdown formats.
+   * Returns line showing more detailed summary of the project, for output in Rich (HTML) or Markdown formats or simple list format.
    * Now uses fontawesome icons for some indicators.
    * @param {string} style
    * @param {boolean} includeFolderName
@@ -1034,8 +1046,7 @@ export class Project {
       }
       output += '\n\t</tr>'
     }
-
-    else if (style === 'Markdown') {
+    else if (style === 'Markdown' || style === 'list') {
       output = '- '
       output += `${this.decoratedProjectTitle(style, includeFolderName)}`
       // logDebug('', `${this.decoratedProjectTitle(style, includeFolderName)}`)
@@ -1055,7 +1066,6 @@ export class Project {
         }
       }
       if (displayProgress && !this.isCompleted && !this.isCancelled) {
-        // const thisPercent = (isNaN(this.percentComplete)) ? '' : ` (${this.percentComplete}%)`
         // Show progress comment if available ...
         if (this.lastProgressComment !== '' && !this.isCompleted && !this.isCancelled) {
           output += `\t${thisPercent} done: ${this.lastProgressComment}`
@@ -1063,8 +1073,6 @@ export class Project {
         // ... else show stats
         else {
           output += `\t${statsProgress}`
-          // Older more detailed stats:
-          // output += `\tc${this.completedTasks.toLocaleString()}${thisPercent} / o${this.openTasks} / w${this.waitingTasks} / f${this.futureTasks}`
         }
       }
       if (displayDates && !this.isPaused && !this.isCompleted && !this.isCancelled) {
@@ -1077,8 +1085,8 @@ export class Project {
             : output
       }
     } else {
-        logWarn('Project::detailedSummaryLine', `Unknown style '${style}'; nothing returned.`)
-        output = ''
+      logWarn('Project::detailedSummaryLine', `Unknown style '${style}'; nothing returned.`)
+      output = ''
     }
     return output
   }