+++ jgclark.Reviews/CHANGELOG.md
@@ -4,7 +4,9 @@ See [website README for more details](https://github.com/NotePlan/plugins/tree/m
 <!-- checkboxes for only display overdue, and showing finished items -->
 
 ## Unreleased
-- jgc: fix to edge case with in getNextNoteToReview() for Dashboard plugin
+- fix to edge case with in getNextNoteToReview() for Dashboard plugin
+- when writing completed and cancelled project summaries to the yearly note, it now puts them as simple list items, not tasks, to avoid a conflict with a Tidy Plugin command.
+- under the hood changes to suit Dashboard 2.0
 
 ## [0.14.0] - 2024-04-???
 ### Added